% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{confint.agreement_fit}
\alias{confint.agreement_fit}
\title{Confidence intervals for an agreement fit}
\usage{
\method{confint}{agreement_fit}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{Object fitted with \code{\link[=agreement]{agreement()}}.}

\item{parm}{Ignored (included for S3 compatibility).}

\item{level}{Confidence level. Default \code{0.95}.}

\item{...}{Ignored.}
}
\value{
For non-inflated data: a list with \code{agreement_est}, \code{agreement_se},
and \code{agreement_ci}. For inflated data: a list with \code{phi_est}, \code{phi_se},
\code{phi_ci}, \code{k0_est}, \code{k0_se}, \code{k0_ci}, \code{k1_est}, \code{k1_se}, \code{k1_ci}.
}
\description{
Confidence intervals for an agreement fit
}
