% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cont2ord}
\alias{cont2ord}
\title{Discretise continuous data}
\usage{
cont2ord(X, K)
}
\arguments{
\item{X}{Vector of continuous data in (0,1).}

\item{K}{Number of ordinal categories.}
}
\value{
Discretised vector
}
\description{
Discretise continuous data
}
