% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_helpers.R
\name{gamma2tau}
\alias{gamma2tau}
\title{Convert gamma to tau using parsimonious parameterization}
\usage{
gamma2tau(GAMMA, K)
}
\arguments{
\item{GAMMA}{Vector of length 2: \link{gamma_1, gamma_2}}

\item{K}{Number of ordinal categories}
}
\value{
Vector of length K+1 with thresholds (tau_0=0, tau_K=1)
}
\description{
Convert gamma to tau using parsimonious parameterization
}
