% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ci}
\alias{get_ci}
\title{Get Agreement confidence interval}
\usage{
get_ci(X, CONFIDENCE = 0.95)
}
\arguments{
\item{X}{Object fitted with \code{\link[=agreement]{agreement()}} function.}

\item{CONFIDENCE}{Confidence level.}
}
\value{
Returns a list with estimate, standard error and confidence interval
}
\description{
Get Agreement confidence interval
}
