% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_helpers.R
\name{get_phi_modified_profile_nested_gamma}
\alias{get_phi_modified_profile_nested_gamma}
\title{Modified profile with gamma parameterization}
\usage{
get_phi_modified_profile_nested_gamma(
  PHI_START,
  GAMMA_START,
  ALPHA_MLE,
  BETA_MLE,
  TAU_MLE,
  PHI_MLE,
  cpp_args = list(),
  lbfgs_control = list(),
  SEARCH_RANGE = 5,
  brent_tol = 1e-04
)
}
\arguments{
\item{PHI_START}{Starting value for PHI}

\item{GAMMA_START}{Starting value for gamma (length 2)}

\item{ALPHA_MLE}{MLE of alpha parameters}

\item{BETA_MLE}{MLE of beta parameters}

\item{TAU_MLE}{MLE of tau parameters}

\item{PHI_MLE}{MLE of phi parameter}

\item{cpp_args}{List of arguments to pass to C++ functions}

\item{lbfgs_control}{List of L-BFGS control parameters}

\item{SEARCH_RANGE}{Search range for PHI optimization}

\item{brent_tol}{Tolerance for Brent's method}
}
\value{
List with optimization results
}
\description{
Modified profile with gamma parameterization
}
