% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_helpers.R
\name{get_phi_profile_nested_gamma}
\alias{get_phi_profile_nested_gamma}
\title{Outer optimization over PHI using nested gamma optimization}
\usage{
get_phi_profile_nested_gamma(
  PHI_START,
  GAMMA_START,
  cpp_args = list(),
  lbfgs_control = list(),
  SEARCH_RANGE = 5,
  brent_tol = 1e-04
)
}
\arguments{
\item{PHI_START}{Starting value for PHI}

\item{GAMMA_START}{Starting value for gamma (length 2)}

\item{cpp_args}{List of arguments to pass to C++ functions}

\item{lbfgs_control}{List of L-BFGS control parameters}

\item{SEARCH_RANGE}{Search range for PHI optimization}

\item{brent_tol}{Tolerance for Brent's method}
}
\value{
List with optimization results
}
\description{
Outer optimization over PHI using nested gamma optimization
}
