% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_range_ll}
\alias{get_range_ll}
\title{Get log-likelihood range}
\usage{
get_range_ll(X, RANGE = 0.2, GRID_LENGTH = 15)
}
\arguments{
\item{X}{Object fitted with \code{\link[=agreement]{agreement()}} function.}

\item{RANGE}{Range around agreement mle.}

\item{GRID_LENGTH}{Number of points to be evaluated within RANGE.}
}
\value{
Return a data.frame with GRID_LENGTH rows and columns
\code{precision}, \code{agreement}, \code{profile} and \code{modified}.
}
\description{
Get log-likelihood range
}
