% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rll}
\alias{get_rll}
\title{Get relative log-likelihood}
\usage{
get_rll(X, RANGE = 0.2, PLOT = TRUE, GRID_LENGTH = 15)
}
\arguments{
\item{X}{Object fitted with \code{agreement()} function}

\item{RANGE}{Range around agreement mle}

\item{PLOT}{Plot relative log-likelihood}

\item{GRID_LENGTH}{Number of points to be evaluated within RANGE.}
}
\value{
Return a data.frame with GRID_LENGTH rows and columns
\code{precision}, \code{agreement}, \code{profile_rll} and \code{modified_rll}.
}
\description{
Get relative log-likelihood
}
