% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_helpers.R
\name{modified_profile_loglik_nested_gamma}
\alias{modified_profile_loglik_nested_gamma}
\title{Modified profile likelihood using gamma parameterization}
\usage{
modified_profile_loglik_nested_gamma(
  RAW_PHI,
  GAMMA_START,
  ALPHA_MLE,
  BETA_MLE,
  TAU_MLE,
  PHI_MLE,
  cpp_args = list(),
  lbfgs_control = list()
)
}
\arguments{
\item{RAW_PHI}{Log of precision parameter}

\item{GAMMA_START}{Starting value for gamma (length 2)}

\item{ALPHA_MLE}{MLE of alpha parameters}

\item{BETA_MLE}{MLE of beta parameters}

\item{TAU_MLE}{MLE of tau parameters}

\item{PHI_MLE}{MLE of phi parameter}

\item{cpp_args}{List of arguments to pass to C++ functions}

\item{lbfgs_control}{List of L-BFGS control parameters}
}
\value{
List with optimization results
}
\description{
Modified profile likelihood using gamma parameterization
}
