% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{par2agr}
\alias{par2agr}
\title{From model parameters to agreement}
\usage{
par2agr(PHI, ALPHA = NULL, BETA = NULL, K0 = NULL, K1 = NULL)
}
\arguments{
\item{PHI}{dispersion parameter}

\item{ALPHA}{item-specific intercepts}

\item{BETA}{worker-specific intercepts}

\item{K0}{zero-inflation threshold}

\item{K1}{one-inflation threshold}
}
\value{
return agreement measure according to the estimated parameters
}
\description{
From model parameters to agreement
}
