% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.rating_data}
\alias{plot.rating_data}
\title{Plot a rating_data object}
\usage{
\method{plot}{rating_data}(x, ...)
}
\arguments{
\item{x}{A \code{rating_data} object from \code{\link[=rating_data]{rating_data()}}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}. Called for its side effect (a rating matrix plot).
}
\description{
Plot a rating_data object
}
\examples{
dt <- sim_data(J = 20, B = 5, AGREEMENT = 0.6,
               ALPHA = rep(0, 20), DATA_TYPE = "continuous", SEED = 1)
rd <- rating_data(dt$rating, dt$id_item, dt$id_worker, VERBOSE = FALSE)
plot(rd)

}
