% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_rll}
\alias{plot_rll}
\title{Plot relative log-likelihood}
\usage{
plot_rll(
  D,
  M_EST = NULL,
  P_EST = NULL,
  M_SE = NULL,
  P_SE = NULL,
  CONFIDENCE = 0.95
)
}
\arguments{
\item{D}{output from \link{get_range_ll}}

\item{M_EST}{agreement estimate from modified profile likelihood}

\item{P_EST}{agreement estimate from profile likelihood}

\item{M_SE}{standard error for agreement estimate from modified profile likelihood}

\item{P_SE}{standard error for agreement estimate from profile likelihood}

\item{CONFIDENCE}{Confidence level to construct confidence intervals}
}
\description{
Plot relative log-likelihood
}
