% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_helpers.R
\name{profile_loglik_nested_gamma}
\alias{profile_loglik_nested_gamma}
\title{Profile likelihood nested optimization using gamma parameterization}
\usage{
profile_loglik_nested_gamma(
  RAW_PHI,
  GAMMA_START,
  cpp_args = list(),
  lbfgs_control = list()
)
}
\arguments{
\item{RAW_PHI}{Log of precision parameter}

\item{GAMMA_START}{Starting value for gamma (length 2)}

\item{cpp_args}{List of arguments to pass to C++ functions}

\item{lbfgs_control}{List of L-BFGS control parameters}
}
\value{
List with optimization results
}
\description{
Profile likelihood nested optimization using gamma parameterization
}
