% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_helpers.R
\name{tau2gamma}
\alias{tau2gamma}
\title{Convert tau to gamma using parsimonious parameterization}
\usage{
tau2gamma(TAU)
}
\arguments{
\item{TAU}{Vector of length K+1 with thresholds}
}
\value{
Vector of length 2: \link{gamma_1, gamma_2}
}
\description{
Convert tau to gamma using parsimonious parameterization
}
