% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cont2ord}
\alias{cont2ord}
\title{Discretise continuous data}
\usage{
cont2ord(X, K, TRESHOLDS = NULL)
}
\arguments{
\item{X}{Vector of continuous data in (0,1).}

\item{K}{Number of ordinal categories.}

\item{TRESHOLDS}{Threshold vector of length K-1. If null, thresholds are assumed to be equispaced.}
}
\value{
Discretised vector
}
\description{
Discretise continuous data
}
