% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lemon}
\alias{lemon}
\title{Squeeze \link{0,100} data}
\usage{
lemon(X, U = NULL)
}
\arguments{
\item{U}{Squeezing parameter}

\item{Y}{Vector of continuous data in \link{0,1}.}
}
\value{
Squeezed vector
}
\description{
Squeeze \link{0,100} data
}
