% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate ordinal or continuous (0,1) ratings}
\usage{
sim_data(
  J,
  B,
  W = J,
  AGREEMENT,
  ALPHA,
  BETA = NULL,
  DATA_TYPE = c("ordinal", "continuous"),
  K = 6,
  SEED = 123
)
}
\arguments{
\item{J}{Number of items.}

\item{B}{Budget per item (i.e. number of workers assigned to each item).}

\item{W}{Maximum number of workers.}

\item{AGREEMENT}{General agreement.}

\item{ALPHA}{Item-specific intercepts.}

\item{BETA}{Worker-specific intercepts.}

\item{DATA_TYPE}{Choose between \code{ordinal} or \code{continuous}.}

\item{K}{Number of categories in case of ordinal data.}

\item{SEED}{RNG seed.}
}
\value{
Returns a dataframe with columns id_items, id_worker and rating
}
\description{
Simulate ordinal or continuous (0,1) ratings
}
